requirements.txt
pyowm
json_database
aiohttp
pydantic
aiofiles
flask
pytz
tzwhere
PyAsyncWeatherAPI